Adding scripts, CSVs, DBs for Uber and Lyft
Set up '/not-website' area of repo to store scripts and data we might share. Following Matt's lead, I uploaded my csv-to-db script (works for lyft dataset), as well as Matt's uber_csv_to_db.py (which was the original work to which I added small bits). CSVs and DBs are here too (although we may want to rethink sharing them this way if they get large).